Refactor out more retail submap code
