+++ src/main/java/legend/lodmod/items/CharmPotionItem.java
@@ -1,12 +1,11 @@
 package legend.lodmod.items;
 
-import legend.game.submap.SMap;
 import legend.game.inventory.Item;
 import legend.game.inventory.UseItemResponse;
+import legend.game.submap.SMap;
 import legend.game.wmap.WMap;
 
 import static legend.game.Scus94491BpeSegment_8004.currentEngineState_8004dd04;
-import static legend.game.Scus94491BpeSegment_800b.hasNoEncounters_800bed58;
 
 public class CharmPotionItem extends Item {
   public CharmPotionItem(final String name, final String description, final String combatDescription, final int price) {
@@ -29,7 +28,7 @@ public class CharmPotionItem extends Item {
       //LAB_80022e40
       response._00 = 8;
       wmap.encounterAccumulator_800c6ae8 = 0;
-    } else if(currentEngineState_8004dd04 instanceof final SMap smap && !hasNoEncounters_800bed58) {
+    } else if(currentEngineState_8004dd04 instanceof final SMap smap && smap.submap.hasEncounters()) {
       response._00 = 8;
       smap.encounterAccumulator_800c6ae8 = 0;
     } else {
